:sparkles: [#15] Ensure the logging formatter can handle exceptions
Request logs without response information are now logged if they have
exception information. The context manager ensures that this exception
information is passed along.

We introspect the request and format it if it's available.

This commit also refactors a bunch of class methods that
don't use anything specific to the instance, into regular
functions. These can be used a bit more broadly to reduce
code duplication in the different formatter paths.